+++ src/main/java/arez/dom/MediaQuery.java
@@ -6,6 +6,7 @@ import arez.annotations.Action;
 import arez.annotations.ArezComponent;
 import arez.annotations.ComputableValueRef;
 import arez.annotations.DepType;
+import arez.annotations.Feature;
 import arez.annotations.Memoize;
 import arez.annotations.Observable;
 import arez.annotations.OnActivate;
@@ -40,7 +41,7 @@ import javax.annotation.Nonnull;
  * }
  * }</pre>
  */
-@ArezComponent
+@ArezComponent( requireId = Feature.DISABLE )
 public abstract class MediaQuery
 {
   @Nonnull